requirements.txt
# ═══════════════════════════════════════════════════════════════
# POLYMARKET ULTIMATE BOT - DEPENDENCIES
# ═══════════════════════════════════════════════════════════════

# Core
python-dotenv>=1.0.0
requests>=2.31.0
websockets>=12.0
aiohttp>=3.9.0

# Data Processing
numpy>=1.26.0
pandas>=2.1.0

# Execution (optional - for live trading)
# py_clob_client>=0.1.0
# eth-account>=0.10.0

# Terminal UI
rich>=13.7.0

# Telegram
# python-telegram-bot>=20.0

# Optional - Sentiment Analysis
# tweepy>=4.14.0
# textblob>=0.17.0

# Optional - Backtesting & Visualization
# matplotlib>=3.8.0
# plotly>=5.18.0